Bugfix/cmakesubmodulefix (#1)
* added picobase64 to .gitmodules
when adding this to my project using externalProject_Add in cmake it gives an error that picobase64 isn't in .gitmodules.

* the joys of copy and paste.
I was too hasty, and too sleep deprived to do it right the first time.

* No root CMakeLists.txt
could have moved it but instead I just added one to the root to include the one in src.

* updated CMakeLists.txt
add src/ to each directory

* Changed to PFL file structure.

PFL As follows: (Note: no .cpp files so no src dir)

build/
A special directory that should not be considered part of the source of the project. Used for storing ephemeral build results. must not be checked into source control. If using source control, must be ignored using source control ignore-lists.

src/
Main compilable source location. must be present for projects with compiled components that do not use submodules.

In the presence of include/, also contains private headers.

include/
Directory for public headers. may be present. may be omitted for projects that do not distinguish between private/public headers. may be omitted for projects that use submodules.

tests/
Directory for tests.

examples/
Directory for samples and examples.

external/
Directory for packages/projects to be used by the project, but not edited as part of the project.

extras/
Directory containing extra/optional submodules for the project.

data/
Directory containing non-source code aspects of the project. This might include graphics and markup files.

tools/
Directory containing development utilities, such as build and refactoring scripts

docs/
Directory for project documentation.

libs/
Directory for main project submodules.

* Works on mac so changed it to OR in cmake